fix memory management of CF objects (fixes `bad_exc_access` crashes)
Change all the uses of CFBridgingRelease to CFAutorelease (or rather the CF_AUTORELEASE macro, that adds a NULL check). Since we never assigned the result of CFBridgingRelease to any variable, ARC was free to release the object at any point it liked. CFBridgingRelease is intended for moving CF objects into the NS world ... not for convenient releasing of CF objects.

Thanks to @muhqu for bringing this to my attention.

// ARC conversion